change hint code
The old hint code is not useful for two reasons. First of all, there is no lexical context, so users who need the hint, have no idea where this block of
code is supposed to fit. There should be a <tbody> tag wrapped around the hint for the lexical context. Secondly, the inner block representing the <td>
tags only has 2 columns of data, which is not what this challenge is asking for - we're asking for 3.
The new hint code provides the needed information to understand what the hint is trying to convey, while still keeping the core aspect of the challenge hidden, and up to the student to figure out.