accept unclosed <td> tags
A regex has been added to detect valid HTML with unclosed <td> tags.
The addedCellsP pattern has been simplified to including only a single <td> tag,
so that now we can get a passing pattern even when <td> tags are left unclosed.

The <td> count is then verified when we do our regex check for valid lines with
unclosed <td>'s.

With this change the issue of valid HTML not passing the grader is being addressed.
Now, <td> tags do not need to be closed in order to pass the challenge.