Fix <thead> pattern
Changed original <thead> pattern to not include linebreaks, for easier
regression testing.

also the test is now under the fail clause, since if the user meets all other requirements, it's not necassary to fail them for changing the thead.
The thead message is only trying to preven users from thinking that the solution for the challenge has anything to do with the thead.